+++ src/test/java/roomescape/reservation/acceptance/ReservationAcceptanceTest.java
@@ -12,6 +12,7 @@ import static roomescape.fixture.TimeFixture.TIME_1;
 
 import io.restassured.RestAssured;
 import io.restassured.http.Cookies;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
@@ -72,7 +73,7 @@ public class ReservationAcceptanceTest {
                 dynamicTest("예약을 추가한다", () -> {
                     Cookies reservationMemberCookies = RestAssuredTemplate.makeUserCookie(MEMBER_BRI);
                     ReservationCreateRequest reservationParams =
-                            new ReservationCreateRequest(date, timeId, themeId, "paymentKey", "orderId", 1000,
+                            new ReservationCreateRequest(date, timeId, themeId, "paymentKey", "orderId", BigDecimal.valueOf(1000),
                                     PaymentType.NORMAL);
                     ReservationResponse response = RestAssuredTemplate.create(reservationParams,
                             reservationMemberCookies);
